requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mysql-connector-python==9.3.0
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.1
PyJWT==2.10.1
rich==13.9.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
